Extract more webdriver methods into specialized mixin classes (#302)
